requirement.txt
fastapi==0.111.0
fastapi-cli==0.0.4
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
orjson==3.10.5
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==3.6
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
